consumer-service/requirements.txt
pika==1.3.2
mysql-connector-python==9.0.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
